Renaming fedshelper file
